requirements.txt
fastapi[all]
python-dotenv==1.0.1